♻️ refactor alertPopup fonction to a new file and call it in FollowersModal
